+++ carousel_parser.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+module CarouselParser
+  autoload :Parser, 'carousel_parser/parser'
+  autoload :ArtworkStruct, 'carousel_parser/artwork_struct'
+end
